+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `OAuth2`, & `SAML` authentication
 - Refactored `Local Auth` to use passport
 - Refactored `auth()` to mix stateless & stateful auth strategies the same time
+- Added `parse()` middleware via `bootstrap()` to coerce `req.body`
 
 ## 0.6.1
 - Upgrading turtle.io to 3.0.15 for etag middleware fix (out of order execution negated it)